refactor: 패키지 interfaces / racing 분리
