+++ src/main/kotlin/camp/nextstep/edu/step3/racing/Car.kt
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.step3
+package camp.nextstep.edu.step3.racing
 
 import java.time.Instant.now
 import kotlin.random.Random
