+++ src/main/kotlin/camp/nextstep/edu/step3/racing/Track.kt
@@ -1,4 +1,4 @@
-package camp.nextstep.edu.step3
+package camp.nextstep.edu.step3.racing
 
 class Track(private val length: Int) {
     private val traces: Array<Trace> = Array(length) { Trace() }
